fix: add support for OMNeT++ 6
https://doc.omnetpp.org/omnetpp/api/APIChanges.html under 5.3, implicit
casting is now preferred over .longValue(). In OMNeT++ 6 .longValue()
will be renamed to .intValue(), but using implicit or static_cast<long>
is sufficient for this.

In SubmoduleIterator, the () operator is deprecated. Use * instead. ()
is removed from the OMNeT++ 6 previews.
Source: https://doc.omnetpp.org/omnetpp/api/classomnetpp_1_1cModule_1_1SubmoduleIterator.html

+++ policies/DIF/EFCP/DTCP/TxControl/TxControlPolicyTCPTahoe/TxControlPolicyTCPTahoe.cc
@@ -40,7 +40,7 @@ void TxControlPolicyTCPTahoe::initialize(int step)
     if(step == 0){
         sigStatTCPTahoeCWND = registerSignal("TCP_Tahoe_CWND");
     }
-    packetSize = par("packetSize").longValue();
+    packetSize = par("packetSize");
     ackPolicy = check_and_cast<SenderAckPolicyTCP *>(getModuleByPath("^.senderAckPolicy"));//^.^.efcp
 }
