+++ policies/DIF/RA/PDUFG/PLQoSAwareMEntries/PLQoSAwareMEntries.cc
@@ -137,7 +137,7 @@ void PLQoSAwareMEntries::onPolicyInit(){
     mType infMetric = par("infinite");
     rt->setInfinite(infMetric);
 
-    maxLat = par("maxLat").longValue();
+    maxLat = par("maxLat");
 
     string myAddr = getModuleByPath("^.^")->par("ipcAddress").stringValue();
