+++ policies/DIF/RA/PDUFG/SimpleLatOrHopMEntries/SimpleLatOrHopMEntries.cc
@@ -98,7 +98,7 @@ void SimpleLatOrHopMEntries::onPolicyInit(){
     mType infMetric = par("infinite");
     rt->setInfinite(infMetric);
 
-    maxLat = par("maxLat").longValue();
+    maxLat = par("maxLat");
 
     string myAddr = getModuleByPath("^.^")->par("ipcAddress").stringValue();
