+++ policies/DIF/Routing/CentralRouting/RoutingManager.cc
@@ -34,10 +34,10 @@ void RoutingManager::onPolicyInit() {
     if(start_Time < 0.0) { start_Time = 0.0; }
     scheduleAt(simTime() + start_Time, start_timer);
 
-    f = par("fabrics").longValue();
-    p = par("pods").longValue();
-    s = par("spines").longValue();
-    t = par("tors").longValue();
+    f = par("fabrics");
+    p = par("pods");
+    s = par("spines");
+    t = par("tors");
 
     current_sent = 0;
     exception_in_seq = 0;
